// CPL/reward/libs/typeorm/src/common/enum.ts
export enum WALLET {
  UNKNOWN = 0,
  EXCHANGE = 1,
  CASHBACK = 2,
  DIVIDEND = 3,
  REWARD = 4,
}

export enum WALLET_LANG {
  UNKNOWN = '',
  EXCHANGE = 'MISSION.MISSION.WALLET_TYPE.EXCHANGE_WALLET',
  CASHBACK = 'MISSION.MISSION.WALLET_TYPE.CASHBACK_WALLET',
  REWARD = 'MISSION.MISSION.WALLET_TYPE.REWARD_WALLET',
}

export enum DELIVERY_METHOD {
  MANUAL = 1,
  AUTO = 2,
}

export enum TARGET_USER {
  USER = 'user',
  REFERRAL_USER = 'referral_user',
}